feat(web): revamp main page and refactor components
