+++ package.json
@@ -10,8 +10,8 @@
   },
   "dependencies": {
     "@fluentui/react-icons": "^2.0.300",
-    "@microsoft/tsdoc": "^0.15.1",
     "@hookform/resolvers": "^5.0.1",
+    "@microsoft/tsdoc": "^0.15.1",
     "@radix-ui/react-accordion": "^1.2.10",
     "@radix-ui/react-alert-dialog": "^1.1.13",
     "@radix-ui/react-aspect-ratio": "^1.1.6",
@@ -22,6 +22,7 @@
     "@radix-ui/react-dialog": "^1.1.13",
     "@radix-ui/react-dropdown-menu": "^2.1.14",
     "@radix-ui/react-hover-card": "^1.1.13",
+    "@radix-ui/react-icons": "^1.3.2",
     "@radix-ui/react-label": "^2.1.6",
     "@radix-ui/react-menubar": "^1.1.14",
     "@radix-ui/react-navigation-menu": "^1.2.12",
@@ -35,6 +36,7 @@
     "@radix-ui/react-slot": "^1.2.2",
     "@radix-ui/react-switch": "^1.2.4",
     "@radix-ui/react-tabs": "^1.1.11",
+    "@radix-ui/react-toast": "^1.2.13",
     "@radix-ui/react-toggle": "^1.1.8",
     "@radix-ui/react-toggle-group": "^1.1.9",
     "@radix-ui/react-tooltip": "^1.2.6",
